remove redundant usecases

+++ app/src/main/java/com/example/currencyexchanger2/currencyexchangescreen/repository/TransactionCountRepository.kt
@@ -1,7 +1,7 @@
 package com.example.currencyexchanger2.currencyexchangescreen.repository
 
 interface TransactionCountRepository {
-    fun getTransactionCount(): Int
+    fun getTransactionCount(): Long
 
     fun incrementTransactionCount()
 }
